+++ sale_order_line_qty_policy_ddt/models/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from . import sale_order
 from . import stock_picking_package_preparation_line
 from . import stock_picking_package_preparation
 from . import account_invoice